+++ src/main/java/edu/mongo10web/service/ProductService.java
@@ -1,6 +1,9 @@
 package edu.mongo10web.service;
 
+import edu.mongo10web.entity.Category;
 import edu.mongo10web.entity.Product;
+import edu.mongo10web.exception.ConflictDataException;
+import edu.mongo10web.exception.InvalidDataFormatException;
 import edu.mongo10web.exception.NotFoundInRepositoryException;
 import edu.mongo10web.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +20,11 @@ public class ProductService {
         this.productRepository = repository;
     }
 
-    public Product create(Product product){
+    public Product create(Product product) {
+        String name = product.getName();
+        if (name==null || name.isBlank()) throw new InvalidDataFormatException();
+        if (product.getCost() == null || product.getCost() <= 0) throw new InvalidDataFormatException();
+        if (!getByName(name).isEmpty()) throw new ConflictDataException();
         return this.productRepository.save(product);
     }
 
@@ -25,20 +32,35 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Product getById(String id){
-        return  productRepository.findById(id).orElseThrow(NotFoundInRepositoryException::new);
+    public Product getById(String id) {
+        if (id==null || id.isBlank()) throw new InvalidDataFormatException();
+        return productRepository.findById(id).orElseThrow(NotFoundInRepositoryException::new);
     }
 
-    public Product update(String id, Product updatedProduct){
+    public List<Product> getByName(String name) {
+        if (name==null || name.isBlank()) throw new InvalidDataFormatException();
+        return productRepository.findByName(name);
+    }
+
+    public Product update(String id, Product updatedProduct) {
+        if (id==null || id.isBlank()) throw new InvalidDataFormatException();
+        if (updatedProduct.getCost() == null || updatedProduct.getCost() <= 0) throw new InvalidDataFormatException();
+        String name = updatedProduct.getName();
+        if (name==null || name.isBlank()) throw new InvalidDataFormatException();
         Product product = getById(id);
-        product.setName(updatedProduct.getName());
+        if(!product.getName().equals(updatedProduct.getName())){
+            List<Product> namedCategory = getByName(name);
+            if(!namedCategory.isEmpty()) throw new ConflictDataException();
+        }
+        product.setName(name);
         product.setCost(updatedProduct.getCost());
         product.setCategory(updatedProduct.getCategory());
         product.setManufacturer(updatedProduct.getManufacturer());
         return productRepository.save(product);
     }
 
-    public void delete(String id){
+    public void delete(String id) {
+        if (id==null || id.isBlank()) throw new InvalidDataFormatException();
         productRepository.deleteById(id);
     }
 }
